Refactor list and join features to hybrid architecture
