+++ app/src/main/java/com/jeong/jjoreum/feature/join/presentation/JoinFormScreen.kt
@@ -1,4 +1,4 @@
-package com.jeong.jjoreum.presentation.ui.join
+package com.jeong.jjoreum.feature.join.presentation
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -29,7 +29,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
 import com.jeong.jjoreum.R
-import com.jeong.jjoreum.presentation.viewmodel.JoinViewModel
+import com.jeong.jjoreum.feature.join.presentation.JoinViewModel
 
 @Composable
 fun JoinFormScreen(
